requirements.txt
discord.py>=2.3.0
pylast>=5.2.0
google-api-python-client>=2.100.0
